requirements.txt
tensorflow
streamlit
opencv-python
pillow
pandas
numpy
uuid